requirements.txt
python-dotenv==1.0.0
mysql-connector-python==8.0.33
discord.py==2.3.2
